+++ CHANGELOG.md
@@ -153,3 +153,9 @@
 
    - Upgrade libyang to v1.0.240
    - Utility function to `show_example_xpaths()`
+ 
+# 0.0.11
+  
+   - remove `data_abstraction_layer`- stublydal is the only option.
+   - add `libyang_gets_xpath` to return generator of DataNode's
+   - add `libyang_get_xpath` to return DataNode (first match or None)